fixing memory issues
