+++ core/amber/src/main/scala/edu/uci/ics/texera/web/service/WorkflowService.scala
@@ -120,5 +120,5 @@ class WorkflowService(wid: String, cleanUpTimeout: Int) extends LazyLogging {
     state.startWorkflow()
   }
 
-  def getJobServiceObservable: Observable[WorkflowJobService] = jobStateSubject
+  def getJobServiceObservable: Observable[WorkflowJobService] = jobStateSubject.onTerminateDetach
 }
